trigger initial job on all schedule methods (#66)
